requirements.txt
face-recognition==1.3.0
numpy==2.2.6
opencv-python==4.12.0.88
Flask==3.1.2
python-dotenv==1.0.0
pillow>=8.0.0
sqlite3